+++ src/components/search-filter/UserSearchFilter.tsx
@@ -1,6 +1,7 @@
 import { UserAvatar } from ":/components/users/avatar/UserAvatar";
 import { SearchFilter } from "./SearchFilter";
 import { UserSearchFilterItem, UserSearchFilterProps } from "./types";
+import { MenuItemBody } from "../menu/MenuItemBody";
 
 export const UserSearchFilter = (props: UserSearchFilterProps) => {
   return (
@@ -8,9 +9,11 @@ export const UserSearchFilter = (props: UserSearchFilterProps) => {
       {...props}
       placeholder={props.placeholder ?? "Search for a name"}
       renderItem={(item) => (
-        <div className="c__search-filter__user-row">
-          <UserAvatar fullName={item.fullName} size="small" />
-          <span>{item.fullName}</span>
+        <div className="c__dropdown-menu-item">
+          <MenuItemBody
+            label={item.fullName}
+            icon={<UserAvatar fullName={item.fullName} size="xsmall" />}
+          />
         </div>
       )}
     />