+++ src/components/search-filter/types.ts
@@ -14,11 +14,13 @@ export type SearchFilterProps<T extends SearchFilterItem = SearchFilterItem> = {
   placeholder?: string;
   items: T[];
   renderItem: (item: T) => ReactNode;
-  onItemSelect?: (item: T) => void;
+  onItemSelect?: (item?: T) => void;
+  selected?: T;
   isLoading?: boolean;
   emptyState?: ReactNode;
   isOpen?: boolean;
   onOpenChange?: (isOpen: boolean) => void;
+  showReset?: boolean;
 };
 
 export type UserSearchFilterItem = SearchFilterItem & {